divided the code into components
